+++ lib/freydis.rb
@@ -1,32 +1,27 @@
+# frozen_string_literal: true
+
 require_relative 'freydis/version'
 require_relative 'freydis/options'
-require_relative 'freydis/init'
-require_relative 'freydis/data'
+require_relative 'freydis/config'
 require_relative 'freydis/disk'
 require_relative 'freydis/disk_luks'
 require_relative 'freydis/cryptsetup'
 require_relative 'freydis/rsync'
-require_relative 'freydis/guard'
 require_relative 'freydis/error'
+require_relative 'freydis/guard'
+require_relative 'freydis/secrets'
 
 module Freydis
+  CONFIG = Config.new
+  CONFIG.load
+  
   class Main
     def initialize(args)
-      @config = args[:config]
-      @cli = args[:cli].options
-      @disk = @cli[:disk]
-
-      Freydis::Guard.disk(@cli[:disk])
+      @argv = args[:argv]
     end
 
     def start
-      init_config
-      encrypt_disk
-      backup
-      restoring
-      opening
-      closing
-      save if @cli[:save]
+      Options.new(@argv)
     end
 
     def bye
@@ -34,67 +29,5 @@ module Freydis
       puts "Bye !"
       exit
     end
-
-    private
-
-    def init_config
-      return unless @cli[:init]
-      Init.run(@cli)
-      save
-    end
-
-    def encrypt_disk
-      return unless @cli[:encrypt]
-      puts "Encrypting disk #{@disk}..."
-      disk = Disk.new(@disk)
-      disk.encrypt(@data)
-    end
-
-    def backup
-      return unless @cli[:backup]
-      raise ArgumentError, "No paths to backup" unless @cli[:paths]
-      raise ArgumentError, "No paths to backup" if @cli[:paths] === []
-
-      puts " ==> Backup on #{@cli[:disk]}..."
-      disk = DiskLuks.new(@cli)
-      disk.open
-      rsync = Rsync.new(@cli)
-      rsync.backup
-      disk.close
-    end
-
-    def restoring
-      return unless @cli[:restore]
-      puts
-      puts " ===> Restoring..."
-      disk = DiskLuks.new(@cli)
-      disk.open
-      rsync = Rsync.new(@cli)
-      rsync.restore
-      disk.close
-    end
-
-    def opening
-      return unless @cli[:open]
-      puts
-      puts " ===> Opening disk #{@disk}."
-      disk = DiskLuks.new(@cli)
-      disk.open
-    end
-
-    def closing
-      return unless @cli[:close]
-      puts
-      puts " ===> Closing disk #{@disk}."
-      disk = DiskLuks.new(@cli)
-      disk.close
-    end
-
-    def save
-      puts
-      puts " ===> Saving options to #{@config}..."
-      Data.new(@config, @cli).save
-    end
   end
 end
-