+++ lib/freydis/disk_luks.rb
@@ -1,37 +1,33 @@
-# lib/disk_luks.rb
+# frozen_string_literal: true
+
+require 'mods/msg'
 
 module Freydis
-  class DiskLuks < Disk
-    def initialize(data)
-      @data = data
-      @disk = data[:disk]
-      if @disk
-        if File.exist? "/dev/disk/by-id/#{@disk}"
-          Freydis::Guard.disk("/dev/disk/by-id/#{@disk}")
-          Freydis::Guard.isLuks("/dev/disk/by-id/#{@disk}")
-        elsif File.exist? "/dev/#{@disk}"
-          Freydis::Guard.disk(@disk)
-          Freydis::Guard.isLuks("/dev/#{@disk}")
-        else
-          puts "#{@disk} value is not supported yet"
-          exit
-        end
-      else
-        puts "No disk."
-        exit 1
-      end
+  module DiskLuks
+    extend Msg
+    module_function
+
+    def encrypt
+      cryptsetup = Freydis::Cryptsetup.new
+      cryptsetup.encrypt
+      cryptsetup.open
+      cryptsetup.format
+      cryptsetup.close
+      success "Disk #{CONFIG.disk} fully encrypted."
     end
 
     def open
-      cryptsetup = Freydis::Cryptsetup.new(@data)
+      cryptsetup = Freydis::Cryptsetup.new
       cryptsetup.close
       cryptsetup.open
       cryptsetup.mount
+      success "Disk #{CONFIG.disk} opened."
     end
 
     def close
-      cryptsetup = Freydis::Cryptsetup.new(@data)
+      cryptsetup = Freydis::Cryptsetup.new
       cryptsetup.close
+      success "Disk #{CONFIG.disk} closed."
     end
   end
 end