+++ lib/freydis/error.rb
@@ -1,10 +1,9 @@
-# lib/error.rb
+# frozen_string_literal: true
 
 module Freydis
-  class InvalidDisk < StandardError
-  end
-  class InvalidLuksDev < StandardError
-  end
-  class InvalidPath < StandardError
-  end
+  class InvalidDisk < StandardError; end
+  class DiskId < StandardError; end
+  class InvalidLuksDev < StandardError; end
+  class InvalidPath < StandardError; end
+  class GPG < StandardError; end
 end