+++ lib/freydis/rsync.rb
@@ -1,53 +1,47 @@
-# lib/rsync.rb
+# frozen_string_literal: true
+
+require 'mods/exec'
 
 module Freydis
   class Rsync
-    def initialize(data)
-      @data = data
-      @mountpoint = '/mnt/freydis'
-      @exclude_paths = [
-        "/dev/*",
-        "/proc/*",
-        "/sys/*",
-        "/tmp/*",
-        "/run/*",
-        "/mnt/*",
-        "/media/*",
-        "/home/*/.thumbnails/*",
-        "/home/*/.cache/mozilla/*",
-        "/home/*/.cache/chromium/*",
-        "/home/*/.local/share/Trash/*",
-        "/lost+found",
+    include Exec
+
+    def initialize
+      @workdir = '/mnt/freydis/backup/'
+      @exclude_paths = %w[
+        /dev/*
+        /proc/*
+        /sys/*
+        /tmp/*
+        /run/*
+        /mnt/*
+        /media/*
+        /var/lib/dhcpcd/*
+        /home/*/.gvfs
+        /home/*/.thumbnails/*
+        /home/*/.cache/*
+        /home/*/.local/share/*
+        /home/*/.Xauthority
+        /home/*/.xsession-errors
+        /lost+found
       ]
-      @opts = "-aAXHvR"
+      @opts = '-aAXHvRx'
     end
 
     def backup
-      add_config
-      exil = @exclude_paths * ","
-      save = @data[:paths] * " "
-      @opts += " --delete"
-      exec("rsync #{@opts} --exclude={#{exil}} #{save} #{@mountpoint}")
+      Freydis::DiskLuks.open
+      mkdir @workdir
+      exil = @exclude_paths * ','
+      save = CONFIG.paths * ' '
+      @opts += ' --delete'
+      x "rsync #{@opts} --exclude={#{exil}} #{save} #{@workdir}"
+      Freydis::DiskLuks.close
     end
 
     def restore
-      exil = @exclude_paths * ","
-      exec("rsync #{@opts} --exclude={#{exil}} #{@mountpoint} /")
-    end
-
-    private
-
-    def add_config
-      if !@data[:paths].include?("#{ENV['HOME']}/.config/freydis")
-        @data[:paths] << "#{ENV['HOME']}/.config/freydis"
-      end
-    end
-
-    def exec(command)
-      sudo = Process.uid != 0 ? 'sudo' : ''
-      if !system("#{sudo} #{command}")
-        raise StandardError, "[-] #{command}"
-      end
+      Freydis::DiskLuks.open
+      x "rsync #{@opts} #{@workdir} /"
+      Freydis::DiskLuks.close
     end
   end
 end